back to no yakmo -- just the Q improvements

+++ extern.pas
@@ -8,7 +8,7 @@ uses
   LazLogger, Windows, Classes, SysUtils, Types, Process, strutils, math;
 
 type
-  TFloat = Single;
+  TFloat = Double;
 
   TFloatDynArray = array of TFloat;
   TFloatDynArray2 = array of TFloatDynArray;
